Move embedding tests into its own file
